캡처 이미지의 전체 크기가 2GB가 넘으면 에러가 나는 문제 해결, deflate압축을 하지 않음으로써 저장 퍼포먼스 향상.

+++ mainwindow.h
@@ -33,6 +33,6 @@ private slots:
     void onFpsChanged(const QString& text);
     void showEditWindow(FileImageStore* store);
     void startRecord();
-    void onUpdateStore(int count, int size);
+    void onUpdateStore(int count, quint64 size);
 };
 #endif // MAINWINDOW_H
